Pi 5/central_code/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ollama==0.4.7
paho-mqtt==2.1.0
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
typing_extensions==4.12.2
